# openapi.yml
openapi: '3.0.2'
info:
  title: simple-twitter
  version: '1.0'
  description: '提供貼文、留言和按讚服務的內容分享平台'
servers:
  - description: '開發伺服器'
    url: http://localhost:3000/api
...
paths:
...
  /tweets:
    get:
      tags:
        - tweets
      operationId: getAllTweet
      summary: 所有貼文
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功獲取貼文資訊、作者資訊、按讚數和留言數
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/tweet'
                    - $ref: '#/components/schemas/tweetNotLiked'
    post:
      tags:
        - tweets
      operationId: addTweet
      summary: 新增貼文
      security:
        - BearerAuth: []
      requestBody:
        description: 發送貼文請求
        content:
          application/json:
            schema:
              type: object
              example: { description: 'this is new post' }
      responses:
        '200':
          description: 成功發送貼文，回傳新增貼文的 id
          content:
            application/json:
              schema:
                type: string
                example: { tweet_id: 100 }
        '400':
          description: 請求錯誤
          content:
            application/json:
              examples:
                emptyContentError:
                  summary: 貼文內容為空
                  value:
                    message: 請填入貼文內容
                tooMuchContentError:
                  summary: 貼文內容過長
                  value:
                    message: 貼文內容不得大於140字
        '507':
          description: 資料庫請求錯誤
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dbError'
  /tweets/{tweet_id}:
    get:
      tags:
        - tweets
      operationId: getTweetById
      summary: 單一貼文
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: tweet_id
          schema:
            type: integer
            default: 2
          required: true
          description: 指定貼文的id
      responses:
        '200':
          description: 成功獲取該貼文的資訊、作者資訊、按讚數和留言
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/tweet'
        '400':
          description: 貼文不存在
          content:
            application/json:
              schema:
                type: object
                example: { message: '貼文不存在' }
  /tweets/{tweet_id}/like:
    post:
      tags:
        - tweets
      operationId: likeTweet
      summary: 對貼文按讚
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: tweet_id
          schema:
            type: integer
            default: 2
          required: true
          description: 欲按讚貼文的id
      responses:
        '200':
          description: 成功按讚
        '400':
          description: 貼文不存在
          content:
            application/json:
              schema:
                type: object
                example: { message: 貼文不存在 }
        '422':
          description: 按讚失敗，已經對該貼文按過讚了
          content:
            application/json:
              schema:
                type: object
                example: { message: 已經對該貼文按過讚了 }
        '507':
          description: 資料庫請求錯誤
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dbError'
  /tweets/{tweet_id}/unlike:
    post:
      tags:
        - tweets
      operationId: unlikeTweet
      summary: 對貼文收回讚
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: tweet_id
          schema:
            type: integer
            default: 2
          required: true
          description: 欲收回讚的貼文id
      responses:
        '200':
          description: 成功對貼文收回讚（硬刪除）
        '400':
          description: 貼文不存在
          content:
            application/json:
              schema:
                type: object
                example: { message: 貼文不存在 }
        '507':
          description: 資料庫請求錯誤
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dbError'
  /tweets/{tweet_id}/replies:
    get:
      tags:
        - tweets
      operationId: getAllReplies
      summary: 指定貼文的留言 & 留言者資訊
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: tweet_id
          schema:
            type: integer
            default: 2
          required: true
          description: 欲取得留言的貼文id
      responses:
        '200':
          description: 獲取該貼文的所有留言和留言者資訊
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: '#/components/schemas/reply'
                    - $ref: '#/components/schemas/reply'
        '400':
          description: 貼文不存在
          content:
            application/json:
              schema:
                type: object
                example: { message: 貼文不存在 }
    post:
      tags:
        - tweets
      operationId: addReply
      summary: 對指定貼文新增留言
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: tweet_id
          schema:
            type: integer
          required: true
          description: 欲新增留言的貼文id
      requestBody:
        description: 留言內容
        content:
          application/json:
            schema:
              type: object
              example: { comment: 'this is reply content' }
      responses:
        '200':
          description: 成功新增留言
        '400':
          description: 留言內容不能為空
          content:
            application/json:
              examples:
                emptyCommentError:
                  summary: 留言內容為空
                  value:
                    message: 請填入留言內容
                unknownTweetError:
                  summary: 貼文不存在
                  value:
                    message: 貼文不存在
  # users
...
  /followships:
    post:
      tags:
        - followships
      operationId: makeFollow
      summary: 追蹤指定帳號
      security:
        - BearerAuth: []
      requestBody:
        description: 欲追蹤帳號的 id
        content:
          application/json:
            schema:
              type: object
              example: { id: 2 }
      responses:
        '200':
          description: 成功追蹤指定帳號
        '400':
          description: 請求錯誤
          content:
            application/json:
              examples:
                accountNotFoundError:
                  summary: 帳號不存在
                  value:
                    message: 追蹤失敗，該帳號不存在
                cantFollowSelfError:
                  summary: 無法追蹤自己
                  value:
                    message: 無法追蹤自己
  /followships/{following_id}:
    delete:
      tags:
        - followships
      operationId: unFollow
      summary: 取消追蹤指定帳號
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: following_id
          schema:
            type: integer
            example: 1
          required: true
          description: 帳號 id
      responses:
        '200':
          description: 成功取消追蹤指定帳號
...
  /admin/tweets/{tweet_id}:
    delete:
      tags:
        - admin
      operationId: deleteTweet
      summary: 管理員刪除指定貼文
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: tweet_id
          schema:
            type: integer
            example: 1
          required: true
          description: 貼文 id
      responses:
        '200':
          description: 成功刪除指定貼文
        '400':
          description: 貼文不存在
          content:
            application/json:
              schema:
                type: object
                example: { message: '刪除失敗，查無該貼文' }
        '401':
          description: 權限不足
          content:
            application/json:
              schema:
                type: object
                example: { message: '權限不足' }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'access token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NvdW50IjoidXNlcjEiLCJ0eXBlIjoiYWNjZXNzIn0.j2QXHeg2HFQNMcmmTeuZjrjSxPe8whGsqrMBjGRz0xg'
  schemas:
...
    tweet:
      properties:
        id:
          type: integer
          example: 2
        description:
          type: string
          example: this is description of tweet
        createdAt:
          type: string
          example: '2023-05-31T03:00:07.000Z'
        updatedAt:
          type: string
          example: '2023-05-31T03:00:07.000Z'
        replyCount:
          type: integer
          example: 1
        likeCount:
          type: integer
          example: 1
        liked:
          type: boolean
          example: 1
        User:
          type: object
          $ref: '#/components/schemas/userInfo'
    tweetNotLiked:
      allOf:
        - $ref: '#/components/schemas/tweet'
      properties:
        liked:
          example: null
...
    userInfo:
      properties:
        id:
          type: integer
          example: 20
        account:
          type: string
          example: user1
        name:
          type: string
          example: 12345678
        avatar:
          type: string
          example: null
...
    reply:
      properties:
        id:
          type: integer
          example: 3
        TweetId:
          type: integer
          example: 1
        comment:
          type: string
          example: this is test comment
        createdAt:
          type: string
          example: '2023-06-01T13:30:00.000Z'
        reply_user:
          type: object
          $ref: '#/components/schemas/userInfo'
...
    dbError:
      properties:
        message:
          type: string
          example: 資料庫請求錯誤